requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-newsletter==0.7
django-pagedown==2.0.3
django-taggit==1.1.0
gunicorn==19.9.0
idna==2.8
ldif3==3.1.1
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
python-card-me==0.9.3
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.25.6
whitenoise==4.1.4
